bug #5/6: fix open modal inside react-router
